Fixed the ticker summary web scraping.

+++ src/main/java/com/trading/tradehub/model/TickerSummaryModel.java
@@ -5,6 +5,6 @@ import java.util.Optional;
 public record TickerSummaryModel(String ticker, String companyName, double marketCap, boolean isOpen,
                                  Optional<Double> currentPrice,
                                  double changeSinceOpen, double changeSinceOpenPercent, double fiftyTwoWeekLow,
-                                 double fiftyTwoWeekHigh, double volume, double averageVolume)
+                                 double fiftyTwoWeekHigh, int volume, int averageVolume)
 {
 }
